Refactor parser specs
